requirements.txt
fastapi~=0.111.1
openai~=1.35.14
pydantic~=2.8.2
python-dotenv~=1.0.1